// payload-types.ts
export interface Weekly {
  id: number;
  title: string;
  summary: string;
  content: string;
  /**
   * 格式：Y + 年份(两位数) + W + 周数(两位数)，例如：Y26W12
   */
  issueNumber: string;
  status: 'draft' | 'published';
  publishDate: string;
  coverImage?: (number | null) | Media;
  links?:
    | {
        label: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  tags?:
    | {
        value: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}